pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "embedsim"
version = "0.1.0"
description = "A Python library for computing Embedding Similarity scores using embeddings"
authors = [
    {name = "Agent CI", email = "hello@agentci.com"},
    {name = "Travis Dent", email = "root@a10k.co"}
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "numpy>=1.21.0",
]

[project.optional-dependencies]
openai = [
    "openai>=1.0.0",
]
sentence-transformers = [
    "sentence-transformers>=2.2.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "mypy",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/embedsim"
Repository = "https://github.com/yourusername/embedsim.git"
Issues = "https://github.com/yourusername/embedsim/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["embedsim*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]